Added spectrogram peak identification, organization
Spectrogram graphing now automatically has green points overlaid on the spectrogram to show the points of peak intensity. The number of points displayed can be modified with a global variable.

Currently, this allows for easier sound visualization. In the future, this will allow for easier data to feed into filters & for machine learning.